requirements.txt
python-monerorpc==0.6.2
qrcode==6.1
Flask==1.1.2
MarkupSafe==1.1.1
requests==2.25.0
gunicorn==20.0.4
Pillow==8.01.1
protobuf==3.15.6
flask_restplus==0.13.0
Werkzeug==0.16.1
PySocks==1.7.1
itsdangerous==2.0.1
